8/19 2:09 챌린지 색 변화

+++ app/src/main/java/com/nyj/routinemaker/Challenge.kt
@@ -10,8 +10,8 @@ data class Challenge(
     var month:String,
     var day:String,
 
-    var percent: Float,
+    var percent: Double,
 
 
 
-)
+    )
